move digit env to within the package.
