--- digit_recognition_environment/__init__.py
@@ -1,1 +1,0 @@
-
