+++ pyproject.toml
@@ -1,9 +1,17 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
 [project]
 name = "r1-vlm"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.12"
+authors = [
+    { name = "Sunil Kumar", email = "sdkumar@hmc.edu" }
+]
+license = { text = "MIT" }
 dependencies = [
     "accelerate>=1.3.0",
     "datasets>=3.2.0",
@@ -27,6 +35,12 @@ dependencies = [
     "flash-attn>=2.7.3",
 ]
 
+[tool.hatch.metadata]
+allow-direct-references = true
+
 [tool.uv.sources]
 trl = { path = "../trl", editable = true }
 verifiers = { path = "../verifiers", editable = true }
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/r1_vlm"]
