+++ src/r1_vlm/environments/simple_vision_env.py
